lib_logrotate: do not let postrotate fool us
Checking for the existence of "rotate" string inside a logrotate entry
used to fool the logic into believing a retention period is defined if
only "postrotate" is defined,
however causing re.sub() call to fail silently, causing a no-op.

We are using re.findall() instead of a substring check now. Also there
are two new unit tests for this condition now.

Works-on: 1969157
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>