Remove cron daily job from root user crontab.
Currently while modifying cron daily job execution time, it is
updated in the root user's crontab (/var/spool/cron/crontabs/root).
This leads to a duplicate execution along with the execution of
cron.daily scripts in /etc/crontab.

This commit aims to manage the execution of the charms' daily cron job
solely from the system crontab, /etc/crontab. It also fixes the existing
unit test `test_cron_daily_schedule` and mocks the crontab instead of
directly writing to it.